Rewrite playerCounts.py using mcstatus
mcstatus is a library that provides objects/methods to query minecraft servers. I've implemented this instead of using `mcapi.us` At the end of the day, I need to make the entire process asynchronous to avoid the 0.5-2 second delay when the script runs.

I had an issue with some of the code in mcstatus, I've commented out a `socket.shutdown()` call. As far as I can tell, the sockets being opened will never have more than one handle (a new one is spawned on each instance of `JavaServer`, so it's A) pointless to have in my use case B) Throwing a benign error for each call, which floods the logs.