https://discord links to discord://discord
Through the use of a custom filter, all instances of `https://discord` will be replaced with `discord://discord`. This will cause these links to open in the Discord App rather than the web browser, if the app is installed.

Turned off autoescaping in the announcements block on homepage, to allow re-formatted `a` element to be displayed after modification.

Also replace all `.gg` links with `.com`, as django's `urlize` filter does not recognize these.