Bot create json, modify embed formatting
The bot will now create the server_announcements.json file if it doen't exist.

Modified the embed formatting to look nicer.

Also removed the update_server_announcement coroutine from the on_raw_message_edit event, as this returns an object that we cannot verify the sender of. This is used for normal announcements as they are all sent from a locked channel.