playerPolls() function call moved to AJAX view
playerPolls() will now be called after a given view is sent to a user. After displaying a spinner, the presented server buttons and counts will display. Utilizing HTMX

- Server Buttons and Server Modals moved to own template from base
- Defined new views under Ajax_Views class, which will return new above templates, as well as a view to return the total count of all servers.
- Added URLS for new views, and added these URLS to robots.txt disallow to prevent them from being crawled.
- HTMX/AJAX request will be triggered for server buttons, modals, and total count on page load, and every 120s thereafter. If a page is loaded while playerCounts.py is still in LOCK time, already present data will be sent. This is nearly immediate.

To put everything into a few words, there is no longer a delay when loading the website if playerPoll needs to run. This is now done with AJAX/AFTER the pagel loads.

- Loaded tail-spin.svg, free from https://samherbert.net/svg-loaders/