Command to get all server announcements
This is meant to be run one, IF messages existed. It will refresh the file with existing mention messages from a server's provided channel. I may put this discord channel ID into a Model so it is passed in automatically.

- Moved code that gets server roles to own function, as it is called twice now.
- Changed existing update_server_announcements to add_server_announcement, to clarify it's function
- New update_server_announcements added, which will get all messages from a servers channel that mention it's role and put it to dictionary. The same way as add does for a specific message that was sent.

To be clear, this is run ONCE. Any new messages will be added to the dictionary. Running this will remove previous additions, and if they were not sent from server's channel , will not re-add them.