/profile htmx GET from defined BASE_USER_PATH
The length of the slice used in user.html from raptormc app to get the authprofiles /profile view is now determined directly from the length of the BASE_USER_URL settings attribute. The user would set this setting to what they are using for their main app 'raptormc''s "user" path.