Add GET parameter to GlobalAnnouncements
This is out of scope for this PR, would have forgotten it though.

Added GET parameter  "amount" to Global_Announcements view for raptorbot. If this parameter is sent in a request, then the view will return an amount of Global Announcements equal to the sent "amount" parameter. This is used currently for the homepage, to only display 5 whereas the Announcements page will display all.