Password reset functional
Not linked to the email backend yet.
Codes are generated when a email link request form is submitted, and saved to the user's "password_reset_token" attribute.

password_reset_confirm view takes the token as a PATH parameter. If the code in PATH matches a user's token, password reset form will be returned. Once password is reset, the code is reset and the user no longer hs the code attribute.

More work to be done, but this is functional