Add de-coupled checks in authprofiles/raptorbot
Similarly to staffapps, these checks will simply be ignored if raptormc is not present in the environment. If it is, it will confirm that a given default page is enabled before giving data back. Otherwise, redirect to 404 or user not found.

In authprofiles, if all_users is disabled, a logged in user will only be able to access their own profile. Attempts to access other user profiles via direct URL will result in a "User not found" page. This does not apply to users that are staff, they can access all profiles via direct URL.