use div for pack name if no picture present
In Edge, the image not found icon will appear with previous code. This is not correct, as the logic was intended to just put the name if the user did not upload an image. Added logic  to put name in a div if server has no picure. Fallback alt attribute still remains on image