Fix deleting RaptorUser
This generally should not be done, but an update to Django made overriding the delete method of a model actually work if you delete from the admin interface, whereas previously you needed to have a receiver for that. After the update, the reciever throws an error as the delete method deletes the associated models for it can.

Moved code to check for proper model from receiver to delete method, and removed receiever.