Add user list informative text
This was a needed informative text in either case, and the fact that it was missing was causing errors with the opengraph implementation.

When I begin adding pages that won't have an informative text and are not model based, I will implement a way to differentiate so that routes.py can know which routes need informative text information for their OG data.