on push to main
This is working as expected now. The main purpose of this is to allow deploying the app easily with Portainer, while also making it more simple when deploying elsewhere. docker-compose-prod.yml is to be used in production, which pulls the built image that is pushed to docker hub on pushes to the main branch, and uses different compose statements to allow this to work properly.

This is in contrast to the existing docker-compose.yml, which reads from the local files and builds a simple image purely containing the dependencies.

The only un-clean thing here is that the web server certificates will need to be swapped out from the docker volume they are placed in, which cannot be done from portainer UI. This is no less clean than it was previously, but I will be looking to have a "production" nginx configuration that will read from default /etc/letsencrypt/live locations. Unsure how this will go currently.