raptormc and gameservers onboarding templates
This is utilizing a modified carousel to display the Network Rules, Server Rules, Banned Items, and Voting Links with a next/previous button. The carousel/card element will remain at 75% of the viewport and force scrolling of the content, so that the next/previous actions do not auto scroll the whole page.