SentEmbedMessage model
This will track the message and channel id of a send server info embed, as well as the server it was giving info for. Will also keep a datetime of the first sent date and the last modified date.

Further, added the update_embeds field to the DiscordBotTasks model for future use.