Delete SentEmbedMessage when deleted from Discord
When a message is deleted from Discord, the ID of tht message will be checked against the SentEmbedMessages in the database.

if there is a a match, the found model will be deleted. This function also sets the "deleted_a_message" attribute of DBInternal model to True, which is used in later code to differentiate between an admin delete and a delete from here.